Some refact&make ci happy
