+++ pallets/math-helper/Cargo.toml
@@ -1,0 +1,48 @@
+[package]
+authors = ['Parallel Team']
+edition = '2021'
+name    = 'pallet-math-helper'
+version = '1.8.5'
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
+
+[dependencies]
+serde               = { version = '1.0.136', features = ['derive'], optional = true }
+codec               = { package = 'parity-scale-codec', version = '3.0.0', features = ['max-encoded-len'], default-features = false }
+frame-support       = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19', default-features = false }
+frame-system        = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19', default-features = false }
+log                 = { version = "0.4", default-features = false }
+num-bigint          = { default-features = false, version = '0.4' }
+num-traits          = { default-features = false, version = '0.2' }
+primitives          = { package = 'parallel-primitives', path = '../../primitives', default-features = false }
+scale-info          = { version = '2.1', default-features = false, features = ['derive'] }
+sp-core             = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19', default-features = false }
+sp-io               = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19', default-features = false }
+sp-runtime          = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19', default-features = false }
+sp-std              = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19', default-features = false }
+substrate-fixed     = { git = "https://github.com/encointer/substrate-fixed", default-features = false }
+
+[dev-dependencies]
+sp-core = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19' }
+
+[features]
+default = ['std']
+std     = [
+  'serde',
+  'codec/std',
+  'frame-support/std',
+  'frame-system/std',
+  'sp-runtime/std',
+  'scale-info/std',
+  'sp-std/std',
+  'primitives/std',
+  'num-bigint/std',
+  'num-traits/std',
+  'sp-core/std',
+]
+
+try-runtime = ['frame-support/try-runtime']
+
+[lib]
+doctest = false
