+++ pallets/traits/Cargo.toml
@@ -25,7 +25,6 @@ xcm           = { git = 'https://github.com/paritytech/polkadot.git', branch = '
 xcm-builder   = { git = 'https://github.com/paritytech/polkadot.git', branch = 'release-v0.9.19', default-features = false }
 xcm-executor  = { git = 'https://github.com/paritytech/polkadot.git', branch = 'release-v0.9.19', default-features = false }
 
-
 [dev-dependencies]
 sp-core = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.19' }
